Update this branch to match CVS head, includes WIN32 improvements.

+++ src/interfaces/libpq/win32.c
@@ -312,7 +312,7 @@ winsock_strerror(int err, char *strerrbuf, size_t buflen)
 	{
 		strerrbuf[buflen - 1] = '\0';
 		offs = strlen(strerrbuf);
-		if (offs > buflen - 64)
+		if (offs > (int)buflen - 64)
 			offs = buflen - 64;
 		sprintf(strerrbuf + offs, " (0x%08X/%lu)", err, err);
 	}
